requirements.txt
streamlit==1.37.0
langchain==0.2.11
langchain-openai==0.1.19
faiss-cpu==1.8.0
sentence-transformers==3.0.1
openai==1.40.0
pypdf==4.2.0
pdfplumber==0.11.4
python-dotenv==1.0.1
tiktoken==0.7.0
numpy==1.26.4
pandas==2.2.2
rapidfuzz==3.9.6
uvicorn==0.30.1
fastapi==0.112.0
pdfplumber
